Fix: Refresh multi-sig extension after enabled permissions
Refresh multi-sig extension after enabled permissions
Fix error toast translation keys
